django_cms/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Image processing
Pillow==10.1.0

# Rich text editor
django-ckeditor==6.7.0

# Admin enhancements
django-admin-sortable2==2.1.10
django-import-export==3.3.3

# Environment variables
python-decouple==3.8

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-slugify==8.0.1
